Cease using deprecated function in json_utils.c
:Release Notes:
The functions in json_utils.h no longer use the deprecated function
json_object_object_get().

:Detailed Notes:
Upstream json-c has deprecated the original method, replacing it with
json_object_object_get_ex() which is NULL safe and guarantees to return
an object pointer on success or NULL on failure.

The JSON utilities have been modified to use the replacement method. As
part of this, boilerplate code in each function - to determine that the
key identified a JVALUE of the correct type - was factored out to a new
local method.

:Testing Performed:
Compile tested only (sleepd is no longer used on any device)

:QA Notes:

:Issues Addressed:
[DRD-3283] Replace cjson with json-c v0.12

Change-Id: I6bfeb82b00ee2229adb9b81d9a332da17c15372e
Reviewed-on: http://gpro.lgsvl.com/136464
Reviewed-by: Martin Jansa <martin.jansa@lge.com>
Tested-by: Martin Jansa <martin.jansa@lge.com>